[info] optimisation spatiale des classes Vector. Dépoussiérage du TGALoader. Amélioration des GetCarteXXXX de Lucy2014. Dépoussiérage du COLLADALoader. Détails : Indentation, NULL->nullptr
Former-commit-id: 93b4511e20b997c328954ec40d2a928e3bdcad7d

+++ Outils/Clock/getMSTime.h
@@ -2,14 +2,13 @@
 
 #include <time.h>
 #ifdef WIN32
-#  include <windows.h>
+# include <windows.h>
 #else
-#  include <sys/time.h>
-#  include <sys/timeb.h>
+# include <sys/time.h>
+# include <sys/timeb.h>
 #endif
 
-inline unsigned int getMSTime()
-{
+inline long getMSTime() {
 #ifdef WIN32
 	return GetTickCount();
 #else
